Changed schemas to lowercase, moved store.ts to src folder.
