+++ client/src/components/pages/Login.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import { useAuthStore } from "../../../store";
+import { useAuthStore } from "../../store";
 
 const Login = () => {
   const login = useAuthStore(state => state.signIn);
