+++ client/src/store.ts
@@ -1,0 +1,11 @@
+import { create } from "zustand";
+
+type UserProfile = {
+  username: string;
+  loggedIn: boolean;
+};
+
+export const useAuthStore = create<UserProfile>(() => ({
+  username: "",
+  loggedIn: false,
+}));
